feat: implement jwt middleware auth
